chore(common): Don't check for conventional commit for fixups
This change allows to add fixup commits by running the command
`git commit --fixup=<sha>` without complaining about the commit
message not being in conventional commit format. (Running the
git command will result in the commit title of the referenced
commit with `fixup!` prepended. It will later be squashed with
the referenced commit in a git rebase.)