feat(developer): cache web debug objects across sessions
Keyman Developer now remembers the 10 most recently debugged keyboards,
models and packages in the web debugger, across sessions. It determines
the recency of use by the last access time of the object, and caches at
most 10 of each type of object when the IDE session ends.

The cache data is stored in %appdata%\Keyman\Keyman Developer\WebDebugCache.json.

The web debugger is now always available when Keyman Developer is
running, even if the user has not yet clicked 'Start Web Debugging', as
otherwise cached web debug objects would not be immediately available.

Keyman Developer will verify that files listed in the cache are still
available, and will remove them on startup if the file disappears.