feat(developer): ngrok configuration
Adds ngrok configuration dialog and processes for downloading, updating
and setting up ngrok.

It is likely this will change with a refactor of the web debugger into a
standalone node module in the future. This will add support for multi-
instance Keyman Developer and be another step towards a cross-platform
IDE.