fix(developer): bundle.sh must succeed
Fixes a build failure after #6074 was merged. The final grep statement
in the line

  cat package.json | grep -v "@keymanapp/models-templates" | grep -v "@keymanapp/models-wordbreakers" > package.json

would fail to find anything to exclude, and return an error value, so
the script would fail, as we use `set -e`. We didn't care that it didn't
find anything to exclude... so we will ignore the return value.