feat(linux): Try to start ibus if not running
When we try to restart ibus and that fails and ibus isn't running
afterwards we directly try to start ibus-daemon.

On Ubuntu 21.10 Impish there are two `ibus-daemon` processes
running for the current user. One of them has the `--xim` flag
and is the one we care about. This change improves the detection
that ibus is running after installing a keyboard.