feat(linux): Try to start ibus after installation
This change deals with restarting ibus in the post-install script.

Also, on Ubuntu 21.10 Impish there are two `ibus-daemon` processes
running for the current user. One of them has the `--xim` flag
and is the one we care about. This change improves the detection
that ibus is running in the post installation script.