feat(linux): Wayland support
The Wayland support of ibus is currently non-working. AFAICT it
only works with XWayland. Also there are still some APIs in Wayland
missing that would allow us to toggle capslock etc, so we use
XWayland if available. If XWayland is not available I guess things
won't completely work, but there's currently not much we can do
about that with ibus.

Fixes #4273.