fix(developer): compile package fileVersion from keyboard FileVersion
Relates to #5992 (part 1 of 2 - part 2 updates the kmn min version in
the compiler).

Updates package fileVersion property to be minimum keyboard version, to
avoid issues with app crashing if functionality is not supported.

As far as I can tell, only Keyman for Windows currently checks the
package fileVersion.

This could be quite nasty, except that the online keyboard repository
build process also extracts the min Keyman version from the keyboard
.kmx and .js, and records it in the .keyboard_info, preventing older
versions of Keyman from seeing unsupported keyboards in the cloud
repository.

Thus, the only gap is for peer distributed keyboards. So we should fix
this gap but it is less urgent than it might otherwise be.