feat(windows): testhost project for debugging keyman32
Scratching an itch. It's very painful to debug Keyman Core on Windows.

This helps a little. It allows you to step through the code without
loading keyman32.dll into every process. It's not perfect, because when
actions are sent, they tend to land in the debugger because SendInput
does not care about the active window! However, if you don't care
about that too much (or if you are willing to try a remote debug setup),
then this should really simplify debugging.