chore(developer): fix issues in package versioning
The package version should always be 7.0 for keyboard versions up to and
including 14.0, because Keyman for Window 14 and earlier only accepted
packages with version number 7.0.

Also, we never stream MinKeymanVersion for each keyboard because that
should be calculated at compile time to ensure that we get correct
version which may otherwise be out of date.