chore(developer): remove Package for Distribution action
Fixes #4168.

This button no longer makes sense when we have template projects, as it
encourages keyboard developers to create multiple packages. The template
project should already have a package.

It is also not hard to create a package if you don't already have one;
this does not reduce the complexity significantly, apart from the mini
feature of adding both the .js and the .kmx in a single step (the .kvk
is added also by virtue of the .kmx being added, but that is done in
any case if you manually add a .kmx to a package).