feat(common): build test keyboards via script
Relates to #5815.

build.sh is a new script that builds any or all of the keyboards in the
common/test/keyboards folder. build.bat is a wrapper for build.sh, just
for Windows CMD.

Assumes that kmcomp.exe has already been built or allows you to specify
your own kmcomp.exe if necessary.