fix(ios): remove busy 100ms waitForFonts timer on iOS
Observed while debugging another issue on iOS. The
VisualKeyboard.waitForFonts function sets up a busy timer to wait for
font loading. This functionality is not even needed on the iOS app,
because it's really only used for the touch alias elements.

However, it was being called, potentially multiple times, leading to the
timer never being removed. On Android, the timer was correctly removed.

This aligns the iOS code path with the Android code path.