fix(web): rely on caller for unmatched keys
Fixes #6005.

If a key has no default behavior, and the expectation is to trigger the
key default, then set the output behavior to null, so that the caller
will do the action.

This fixes a regression in #5553, which handled the situation where a
rule was triggered but then default output was expected (common example
is final sigma transform only when a word is finished).

The regression resulted in suggestions rotating unexpectedly when
pressing a modifier key.