fix(windows): add install defaults flag to cmd line
The install defaults flag will set the registry values
to their default values.

This flag can also be used by the installer itself when it needs to
do a restart to continue installing. It can check to see if this is
a fresh install or an upgrade and set the flag accordingly.