fix(developer): kmcomp package compiler race condition
Fixes a race condition when compiling multiple packages, where the
temp path may be re-used before the compiler has finished with it.

To take advantage of this fix, I wanted to make sure I was using the
right kmcomp.exe when building unit tests for Keyman Core. This
uncovered a couple of other things:

* Updates Keyman Core build infrastructure to use the current kmcomp.exe
  rather than a system-installed one.
* Corrects a bug that was identified by the 15.0 version of kmcomp in
  one unit test (duplicate `&NAME` stores -- the meson test script
  prepends a `&NAME` store to each source file during build).

@keymanapp-test-bot skip