chore(windows): extract TJCLMapScanner down to separate unit
TJCLMapScanner was the only class used in the entire JCL, in a
build-time project (sentrytool). For now, I have copied this class plus
its handful of dependent functions into Keyman.System.JclMapScanner so
that we can remove 3381 files from the repository.

We do not plan to use JCL in the future.