chore(web): add 'index.html' to parent links in test pages
We need a '../index.html' link rather than '../' on the three
second-level KeymanWeb test pages, in order to get to the index page for
artifacts when the test pages are hosted on build.palaso.org --
otherwise TeamCity returns a 404. (This is probably also good for
file-system-based navigation.)

We don't need the same reference on other directories, just for the
top-level index.html.

Identified at https://github.com/keymanapp/keyman/pull/6557#issuecomment-1157526326